requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==7.0.0
distro==1.9.0
docx==0.2.4
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.15.4
fitz==0.0.1.dev2
flask==3.0.3
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.0
idna==3.7
importlib-metadata==8.0.0
importlib-resources==6.4.0
isodate==0.6.1
itsdangerous==2.2.0
jinja2==3.1.4
looseversion==1.3.0
lxml==5.2.2
MarkupSafe==2.1.5
networkx==3.1
nibabel==5.2.1
nipype==1.8.6
numpy==1.24.4
openai==1.36.1
packaging==24.1
pandas==2.0.3
pathlib==1.0.1
pillow==10.4.0
prov==2.0.1
pydantic==2.8.2
pydantic-core==2.20.1
pydot==3.0.1
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==7.0.0
regex==2024.5.15
requests==2.32.3
safetensors==0.4.3
scipy==1.10.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
tokenizers==0.19.1
tqdm==4.66.4
traits==6.3.2
transformers==4.42.4
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
werkzeug==3.0.3
zipp==3.19.2
